apollo_l1_events: move transaction_manager tests to a separate file and simplify metric export
Goal: address the CLAUDE: review comments left on the oldest-pending-L1-timestamp
metric branch.

Changes:
- oldest pending tx timestamp now uses .map(|t| t.0).unwrap_or_default() instead of
  map_or(0, ...).
- Moved the transaction_manager unit tests out of transaction_manager.rs into a sibling
  transaction_manager_tests.rs, wired via #[cfg(test)] #[path = ...] mod, matching the
  crate's existing l1_scraper_tests.rs / l1_events_provider_tests.rs convention.

Decision points:
- The map/unwrap_or_default form follows the inline CLAUDE comment, though it runs
  counter to code-style.md's preference for map_or; applied as the comment directed.
