+++ crates/apollo_l1_events/src/transaction_manager.rs
@@ -22,6 +22,10 @@ use crate::transaction_record::{
     TransactionState,
 };
 
+#[cfg(test)]
+#[path = "transaction_manager_tests.rs"]
+mod transaction_manager_tests;
+
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct TransactionManager {
     /// Storage of all l1 handler transactions --- keeps transactions until they can be safely
@@ -363,12 +367,7 @@ impl TransactionManager {
         self.proposable_index
             .values()
             .flatten()
-            .filter_map(|tx_hash| match self.records.get(tx_hash)?.tx {
-                TransactionPayload::Full { created_at_block_timestamp, .. } => {
-                    Some(created_at_block_timestamp)
-                }
-                TransactionPayload::HashOnly(_) => None,
-            })
+            .filter_map(|tx_hash| self.records.get(tx_hash)?.tx.created_at_block_timestamp())
             .min()
     }
 
@@ -414,8 +413,10 @@ impl TransactionManager {
             // timestamp (not the age) so the alert can compute `time() - timestamp` and have the
             // age grow correctly even while the provider is idle. 0 means nothing is
             // pending.
-            let oldest_pending_tx_timestamp =
-                self.oldest_pending_l1_block_timestamp().map_or(0, |timestamp| timestamp.0);
+            let oldest_pending_tx_timestamp = self
+                .oldest_pending_l1_block_timestamp()
+                .map(|timestamp| timestamp.0)
+                .unwrap_or_default();
             L1_MESSAGE_PROVIDER_OLDEST_PENDING_TX_L1_TIMESTAMP_SECONDS
                 .set_lossy(oldest_pending_tx_timestamp);
 
@@ -522,66 +523,5 @@ impl Sub<u128> for StagingEpoch {
 
     fn sub(self, rhs: u128) -> Self::Output {
         Self(self.0 - rhs)
-    }
-}
-
-#[cfg(test)]
-mod transaction_manager_tests {
-    use indexmap::IndexMap;
-
-    use super::*;
-    use crate::test_utils::l1_handler;
-
-    /// Builds a `Pending`, full-payload record. The L1 block timestamp (what the metric reports) is
-    /// set independently of the scrape timestamp so tests can distinguish the two.
-    fn pending_record(
-        tx_hash: usize,
-        l1_block_timestamp: u64,
-        scrape_timestamp: UnixTimestamp,
-    ) -> (TransactionHash, UnixTimestamp, TransactionRecord) {
-        let tx = l1_handler(tx_hash);
-        let hash = tx.tx_hash;
-        let record = TransactionRecord::new(TransactionPayload::Full {
-            tx,
-            created_at_block_timestamp: BlockTimestamp(l1_block_timestamp),
-            scrape_timestamp,
-        });
-        (hash, scrape_timestamp, record)
-    }
-
-    fn manager_with_pending(
-        pending: Vec<(TransactionHash, UnixTimestamp, TransactionRecord)>,
-    ) -> TransactionManager {
-        let mut records = IndexMap::new();
-        let mut proposable_index: BTreeMap<UnixTimestamp, Vec<TransactionHash>> = BTreeMap::new();
-        for (hash, scrape_timestamp, record) in pending {
-            proposable_index.entry(scrape_timestamp).or_default().push(hash);
-            records.insert(hash, record);
-        }
-        TransactionManager::create_for_testing(
-            records.into(),
-            proposable_index,
-            StagingEpoch::new(),
-            TransactionManagerConfig::default(),
-            BTreeMap::new(),
-        )
-    }
-
-    #[test]
-    fn oldest_pending_l1_block_timestamp_returns_min_block_timestamp() {
-        // Scrape order (proposable index key) deliberately differs from L1 block order: the tx
-        // scraped first (key 10) has the newest L1 timestamp, so a "first entry" implementation
-        // would report the wrong value.
-        let manager = manager_with_pending(vec![
-            pending_record(1, 300, 10),
-            pending_record(2, 100, 20),
-            pending_record(3, 200, 30),
-        ]);
-        assert_eq!(manager.oldest_pending_l1_block_timestamp(), Some(BlockTimestamp(100)));
-    }
-
-    #[test]
-    fn oldest_pending_l1_block_timestamp_is_none_when_nothing_pending() {
-        assert_eq!(manager_with_pending(vec![]).oldest_pending_l1_block_timestamp(), None);
     }
 }
